+++ python/benchmarks/bench_contact_3D.py
@@ -216,7 +216,7 @@ def demo_stacked_cubes(theta, ct, noslip, num_refinements, N0, timings=False):
     bcs = [bc_bottom, bc_top]
 
     # Elasticity parameters
-    E = 1.0e3
+    E = PETSc.ScalarType(1.0e3)
     nu = 0
     mu = dolfinx.Constant(mesh, E / (2.0 * (1.0 + nu)))
     lmbda = dolfinx.Constant(mesh, E * nu / ((1.0 + nu) * (1.0 - 2.0 * nu)))
@@ -229,7 +229,7 @@ def demo_stacked_cubes(theta, ct, noslip, num_refinements, N0, timings=False):
     u = ufl.TrialFunction(V)
     v = ufl.TestFunction(V)
     a = ufl.inner(sigma(u), ufl.grad(v)) * ufl.dx
-    rhs = ufl.inner(dolfinx.Constant(mesh, (0, 0, 0)), v) * ufl.dx
+    rhs = ufl.inner(dolfinx.Constant(mesh, PETSc.ScalarType((0, 0, 0))), v) * ufl.dx
 
     dolfinx_mpc.utils.log_info("Create constraints")